Remove rediness probe from example.
Probe does not interact well with Kubernetes as configured (with one service per node).  Until a ZK node is ready, it won't be added to its service.  That means other ZK nodes won't be able to reach it by making a connection to it.  The ZK node could make a connection to other nodes if they are ready, but a ZK quick makes that not possible.  ZK will drop an incoming connection from a peer if the peer's server id is greater.  It wants to make sure that connections between ZK nodes are made from lower numbered servers to higher ones.